Thêm button tải chart, điều chỉnh Order Entity, OrderLines Entity cho đồng bộ với ôg Hải

+++ src/main/java/hcmute/vn/springonetomany/Model/ProfitReportByDay.java
@@ -13,7 +13,7 @@ import lombok.Setter;
 @Setter
 public class ProfitReportByDay {
 	private String productName;
-	private Integer profit;
+	private Double profit;
